Factor benchmarks into separate files
Makes the benchmarking suite a little more manageable
