+++ Cargo.toml
@@ -9,6 +9,7 @@ repository  = "https://github.com/tendermint/signatory/tree/master"
 readme      = "README.md"
 categories  = ["authentication", "cryptography", "no-std"]
 keywords    = ["crypto", "cryptography", "no_std", "security", "signatures"]
+autobenches = false
 
 [badges]
 circle-ci = { repository = "tendermint/signatory" }
@@ -19,7 +20,7 @@ ed25519-dalek = { version = "0.6", optional = true, default-features = false, fe
 generic-array = { version = "0.9", optional = true }
 lazy_static = { version = "1", optional = true }
 rand = { version = "0.5", optional = true }
-ring = { version = "0.13.0", optional = true }
+ring = { version = "0.13", optional = true }
 secp256k1 = { version = "0.9", optional = true }
 sha2 = { version = "0.7", optional = true }
 sodiumoxide = { version = "0.1", optional = true }
@@ -69,5 +70,13 @@ debug-assertions = false
 codegen-units = 1
 
 [[bench]]
-name = "ed25519"
+name = "dalek"
+harness = false
+
+[[bench]]
+name = "ring"
+harness = false
+
+[[bench]]
+name = "sodiumoxide"
 harness = false
